Fix 403 error - allow external users to exclude sites
Critical bug fix: External users MUST be able to set sites as excluded.
This is a core feature of the review process - clients need to be able
to choose which sites they want (included) and don't want (excluded).

The previous restriction was completely wrong and broke the fundamental
purpose of the review page.

Changes:
- Remove incorrect restriction blocking external users from excluding sites
- Keep restriction only on setting exclusionReason (internal notes)
- External users can now properly set: included/excluded/saved_for_later